backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "evan chee", email = "chee@wustl.edu"},
]
dependencies = ["fastapi[standard]>=0.115.8", "rasterio>=1.4.3", "python-dotenv>=1.0.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
